refactor(observability): move the turn saga out of core
Bus subscriptions belong in a sagas directory, not core/. The counter
instrument stays in core as a plain helper; the subscription moves to the
cross-context sagas directory, alongside the agent-artifacts sweeper.

Also drops a doc claim that the counter and the activity log cannot
disagree — each is separately enabled and each fails on its own, so a
divergence means a sink is off or dropping, not two notions of a turn.
