+++ packages/api-server/src/bootstrap.ts
@@ -79,7 +79,7 @@ import { composeE2eModule } from "./modules/e2e/compose.js";
 import { composeTermsModule } from "./modules/terms/index.js";
 import { loadConfig } from "./config.js";
 import { configureLogger, getLogger } from "./core/logger.js";
-import { startTurnMetricsSaga } from "./core/turn-metrics.js";
+import { startTurnMetricsSaga } from "./sagas/turn-metrics.js";
 import { formatError } from "./core/format-error.js";
 import type { ApiServerDeps } from "./apps/api-server/deps.js";
 import {
